back end work in user recipes.js, addition of cpntrollers folder with API.js moved into here/ recipesController added.
