+++ models/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+    userRecipes: require("./userRecipes"),
+    User: require("./User"),
+}
+
+// API.JS MIGHT HAVE TO MOVE TO FRONT-END UTILS FOLDER TO MAKE THE CALL
